Update NIPM Library/List Upgradable Packages.vi
Modified "List Upgradable Packages.vi" so that it outputs a 2D string array (table) that now lists not only the upgradable package names but also the package versions to update to.  This now matches datatype of related list packages VIs.
Note: This is a breaking change because the data type changed from 1D array to 2D array.

Added detailed context help for VI description and input/output descriptions for all VIs in this API.